Pass an array reference to Optimizer
Previously, the optimizer received a copy of the original Node array.
This resulted in changes inside the optimizer not being used outside of
that class.

This commit fixes the problem by removing the LayoutNode type and
merging it with the CleanNode type. This allows the optimizer properties
to directly re-use the TheoryGraph array.